src/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.20
pypdf2==3.0.1
python-docx==1.1.0
sentence-transformers==2.2.2
chromadb==0.4.22
python-multipart==0.0.9
pydantic-settings==2.0.3
pymupdf==1.24.9
motor==3.6.0
pymongo<5
openai==1.35.13
cohere==5.5.8
qdrant-client==1.15.1
# Monitoring and metrics
prometheus-client==0.21.1
starlette-exporter==0.23.0
fastapi-health==0.4.0